fix compilation with latest V after https://github.com/vlang/v/pull/23061

+++ 2024/01/drakeerv.v
@@ -5,14 +5,14 @@ import arrays
 fn main() {
 	lines := os.read_file('locations.input')!.split_into_lines()
 
-	mut first_numbers := []u32{}
-	mut second_numbers := []u32{}
+	mut first_numbers := []int{}
+	mut second_numbers := []int{}
 
 	for i in 0 .. lines.len {
 		numbers := lines[i].split('   ')
 
-		first_numbers << u32(strconv.parse_uint(numbers[0], 10, 32)!)
-		second_numbers << u32(strconv.parse_uint(numbers[1], 10, 32)!)
+		first_numbers << int(strconv.parse_uint(numbers[0], 10, 32)!)
+		second_numbers << int(strconv.parse_uint(numbers[1], 10, 32)!)
 	}
 
 	first_numbers.sort()
@@ -21,7 +21,7 @@ fn main() {
 	mut total := 0
 
 	for i in 0 .. lines.len {
-		mut distance := i32(first_numbers[i] - second_numbers[i])
+		mut distance := first_numbers[i] - second_numbers[i]
 		if distance < 0 {
 			distance = -distance
 		}
